refactored parsing
