+++ digitalpy/core/parsing/format.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
-from digitalpy.routing.request import Request
-from digitalpy.routing.response import Response
+from digitalpy.core.zmanager.request import Request
+from digitalpy.core.zmanager.response import Response
 
 class Format(ABC):
